second commit

+++ src/main/java/com/sample/demo/repositories/ChildRepository.java
@@ -3,5 +3,5 @@ package com.sample.demo.repositories;
 import com.sample.demo.domains.Child;
 import org.springframework.data.repository.CrudRepository;
 
-public interface ChildRepository extends CrudRepository<Child, Long> {
+public interface ChildRepository extends CrudRepository<Child, Integer> {
 }
